requirements.txt
flask
openpyxl
openai
python-dotenv
pandas
requests
psycopg2-binary
